Refactoring to fix issues
