+++ scapy/arch/bpf/__init__.py
@@ -1,0 +1,5 @@
+# Guillaume Valadon <guillaume@valadon.net>
+
+"""
+Scapy *BSD native support
+"""
