+++ scapy/sendrecv.py
@@ -128,7 +128,7 @@ def sndrcv(pks, pkt, timeout = None, inter = 0, verbose=None, chainCC=0, retry=0
                                     break
                             r = None
                             if conf.use_bpf:
-                                from scapy.arch.bpf import bpf_select
+                                from scapy.arch.bpf.supersocket import bpf_select
                                 inp = bpf_select(inmask)
                                 if pks in inp:
                                     r = pks.recv()
@@ -494,7 +494,7 @@ def sndrcvflood(pks, pkt, prn=lambda (s,r):r.summary(), chainCC=0, store=1, uniq
     try:
         while 1:
             if conf.use_bpf:
-                from scapy.arch.bpf import bpf_select
+                from scapy.arch.bpf.supersocket import bpf_select
                 readyr = bpf_select([rsock])
                 _, readys, _ = select([], [ssock], [])
             else:
@@ -617,7 +617,7 @@ interfaces)
                 if remain <= 0:
                     break
             if conf.use_bpf:
-                from scapy.arch.bpf import bpf_select
+                from scapy.arch.bpf.supersocket import bpf_select
                 ins = bpf_select(sniff_sockets, remain)
             else:
                 ins, _, _ = select(sniff_sockets, [], [], remain)
@@ -694,7 +694,7 @@ stop_filter: python function applied to each packet to determine
                 if remain <= 0:
                     break
             if conf.use_bpf:
-                from scapy.arch.bpf import bpf_select
+                from scapy.arch.bpf.supersocket import bpf_select
                 ins = bpf_select([s1, s2], remain)
             else:
                 ins, _, _ = select([s1, s2], [], [], remain)
